\name{partialCI-package}
\alias{partialCI-package}
\alias{partialCI}
\docType{package}
\title{
\packageTitle{partialCI} Partial Cointegration Models
}
\description{
A collection of time series is said to be partially cointegrated if they have a linear combination that is partially autoregressive, e.g., that can be represented as a sum of an autoregressive series without unit roots and a random walk.  This may be useful in modeling certain sets of financial time series.
}
\keyword{ ts }
\keyword{ models }% __ONLY ONE__ keyword per line
